Move DifficultyObject to preprocessing package
