+++ app/oppai/aim.go
@@ -3,6 +3,7 @@ package oppai
 import (
 	"github.com/wieku/danser-go/app/beatmap/difficulty"
 	"github.com/wieku/danser-go/app/beatmap/objects"
+	"github.com/wieku/danser-go/app/oppai/preprocessing"
 	"math"
 )
 
@@ -21,7 +22,7 @@ func NewAimSkill(useFixedCalculations bool, d *difficulty.Difficulty) *Skill {
 	return skill
 }
 
-func aimStrainValue(skill *Skill, current *DifficultyObject) float64 {
+func aimStrainValue(skill *Skill, current *preprocessing.DifficultyObject) float64 {
 	if _, ok := current.BaseObject.(*objects.Spinner); ok {
 		return 0
 	}
